consolidate app proto conversions
